FIX tests : Add test_streamlit in pytest MOD workflow/codecov : set pytest in venv
