+++ .github/workflows/codecov.yml
@@ -12,24 +12,24 @@ jobs:
         python-version: ["3.12"]
     steps:
       - uses: actions/checkout@v4
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@master
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v7
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
-        run: |
-          git submodule update --init --remote
-          pip install pytest pytest-cov setuptools
-          python -m pip install .
-          PWD=$(pwd)
-          export PYTHONPATH=$PWD/src:$PWD/lib/emhass/src/:$PYTHONPATH
-          echo "PYTHONPATH=$PYTHONPATH"
-          echo "PYTHONPATH=$PYTHONPATH" >> $GITHUB_ENV
+          enable-cache: true
+          cache-dependency-glob: "uv.lock"
+
+      - name: Sync dependencies
+        run: uv sync --frozen
 
       - name: Run tests
-        run: pytest --cov --cov-report=xml --ignore=tests/test_emhass_adapter.py --ignore=src/openhems/emhass
+        run: uv run pytest --cov --cov-report=xml \
+                --ignore=tests/test_emhass_adapter.py \
+                --ignore=src/openhems/emhass
 
       - name: Upload results to Codecov
         uses: codecov/codecov-action@v4
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
+
