+++ pyproject.toml
@@ -35,7 +35,7 @@ dependencies = [
 	'packaging',
 	'astral',
 	'GitPython',
-	'toml',
+	"toml",
 	'scikit-opt',
 	# For genetic algorithm
 	'scikit-opt',
@@ -53,6 +53,8 @@ dependencies = [
 	'pulp',
 	"streamlit>=1.57.0",
 	"streamlit-monaco-yaml>=0.2.0",
+	"pytest>=9.0.3",
+	"pytest-cov>=7.1.0",
 ]
 license = { file = "LICENSE", "expression" = "GPL-3.0-or-later" }
 # license = "GPL-3.0-or-later"
